Move the global hash map to its own module
